Refactor kubeconfig management and simplify code structure
Moved kubeconfig handling utilities to `pkg/k8s` for better modularization. Simplified the structure and reduced redundant logic, improving maintainability and readability.
